homework5.1

+++ src/services/contacts.js
@@ -3,14 +3,14 @@ import { calculatePaginationData } from "../utils/calculatePaginationData.js";
 import { SORT_ORDER } from "../constants/index.js"
 
 export const getAllContact = async ({ page, perPage, sortOrder = SORT_ORDER.ASC,
-    sortBy = '_id', filter = {} }) => {
+    sortBy = '_id', filter = {}, userId }) => {
     const limit = perPage;
     const skip = (page - 1) * perPage;
+    const userFilter = { ...filter, userId };
 
-    const contactQuery = ContactsCollection.find(filter);
+    const contactQuery = ContactsCollection.find(userFilter);
 
-    const contactCount = await ContactsCollection.find().merge(contactQuery).countDocuments(filter);
-    console.log(contactCount);
+    const contactCount = await ContactsCollection.find().merge(contactQuery).countDocuments(userFilter);
 
     const contacts = await contactQuery.skip(skip).limit(limit).sort({ [sortBy]: sortOrder }).exec();
     const paginationData = calculatePaginationData(contactCount, perPage, page);
@@ -21,8 +21,8 @@ export const getAllContact = async ({ page, perPage, sortOrder = SORT_ORDER.ASC,
     };
 };
 
-export const getContactById = async (contactId) => {
-    const contact = await ContactsCollection.findById(contactId);
+export const getContactById = async (contactId, userId) => {
+    const contact = await ContactsCollection.findById({ _id: contactId, userId });
 
     return contact;
 };
@@ -33,9 +33,9 @@ export const createContact = async (payload) => {
     return contact;
 };
 
-export const updateContact = async (contactId, payload, options = {}) => {
+export const updateContact = async (contactId, userId, payload, options = {}) => {
     const rawResult = await ContactsCollection.findOneAndUpdate(
-        { _id: contactId },
+        { _id: contactId, userId },
         payload,
         {
             new: true,
@@ -52,8 +52,8 @@ export const updateContact = async (contactId, payload, options = {}) => {
     };
 };
 
-export const deleteContact = async (contactId) => {
-    const contact = await ContactsCollection.findByIdAndDelete({ _id: contactId });
+export const deleteContact = async (contactId, userId) => {
+    const contact = await ContactsCollection.findByIdAndDelete({ _id: contactId, userId },);
 
     return contact;
 };
